Add "tini" to dockerfile (#88)
Tini is a tiny init process for docker containers that ensure that they handle SIGTERM/SIGKILL etc. in all environments correctly. You can read more here: https://github.com/krallin/tini

This is added as an `ENTRYPOINT`, so that it will automatically be used for any container built on top of this one which overrides `CMD`. Since TINI is an init process, adding it as an entrypoint should be completely backwards compatible with any already working container build on top of this one.